backend/requirements.txt
autopep8==1.6.0
click==8.1.2
colorama==0.4.4
distlib==0.3.6
filelock==3.11.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
platformdirs==3.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.6
pycodestyle==2.8.0
python-decouple==3.6
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==2.0.9
toml==0.10.2
typing_extensions==4.5.0
virtualenv==20.21.0
Werkzeug==2.2.3
zipp==3.8.0
